create new applications + new tables
