+++ oc_lettings_site/views.py
@@ -1,8 +1,6 @@
 from django.shortcuts import render
-from .models import Letting, Profile
-
-
-
+from lettings.models import Letting
+from profiles.models import Profile
 
 # Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque molestie quam lobortis leo consectetur ullamcorper non id est. Praesent dictum, nulla eget feugiat sagittis, sem mi convallis eros,
 # vitae dapibus nisi lorem dapibus sem. Maecenas pharetra purus ipsum, eget consequat ipsum lobortis quis. Phasellus eleifend ex auctor venenatis tempus.
